+++ core/class/jeeHistoGraph.class.php
@@ -92,7 +92,20 @@ class jeeHistoGraph extends eqLogic {
     foreach ($configs as $key) {
         $this ->setConfiguration($key[0], $key[1]);
     }
+    $actualVersion = config::byKey('version', 'jeeHistoGraph', '0.0', true);
+    $this   ->setConfiguration('version', $actualVersion);
     $this->save();
+    $refresh = $this->getCmd('action', 'refresh');
+    if (!is_object($refresh)) {
+        log::add("jeeHistoGraph", "debug", "création de refresh");
+        $refresh = new jeeHistoGraphCmd();
+        $refresh->setName(__('Rafraichir', __FILE__));
+    }
+    $refresh->setEqLogic_id($this->getId());
+    $refresh->setLogicalId('refresh');
+    $refresh->setType('action');
+    $refresh->setSubType('other');
+    $refresh->save();    
   }
 
   // Fonction exécutée automatiquement avant la mise à jour de l'équipement
@@ -169,7 +182,14 @@ public static function config() {
         $config[] = ["graph{$g}_show_refPrec", 1];
         $config[] = ["graph{$g}_inverted", 1];
         $config[] = ["graph{$g}_tooltip_enabled", 1];
-        $config[] = ["graph{$g}_dataLabels_overlaps"];
+        $config[] = ["graph{$g}_dataLabels_overlaps", 0];
+        $config[] = ["graph{$g}_3D_enabled", 0];
+        $config[] = ["graph{$g}_3D_alpha", 15];
+        $config[] = ["graph{$g}_3D_beta", 15];
+        $config[] = ["graph{$g}_3D_depth", 25];
+        $config[] = ["graph{$g}_3D_view_distance", 0];
+        $config[] = ["graph{$g}_zoom_axe_x", 1];
+        $config[] = ["graph{$g}_zoom_axe_y", 1];
         for ($i = 1; $i <= 10; $i++) {
             $index = str_pad($i, 2, '0', STR_PAD_LEFT);
             $config[] = ["graph{$g}_index{$index}_nom", ''];
@@ -234,10 +254,33 @@ public function toHtml($_version = 'dashboard', $eqLogic = null) {
         $inverted = 'false';
         $tooltipEnabled = 'true';
         $xRangeSelectorButtonPosition = 0;
+        $yRangeSelectorButtonPosition = $showTitle ? -30 : 0;
 
         $configNavigatorEnabled = $eqLogic->getConfiguration("graph{$g}_navigator", 0) ? 'true' : 'false';
         $configBarreEnabled = $eqLogic->getConfiguration("graph{$g}_barre", 0) ? 'true' : 'false';
         $configButtonsEnabled = $eqLogic->getConfiguration("graph{$g}_buttons", 0) ? 'true' : 'false';
+        $config3DEnabled = $eqLogic->getConfiguration("graph{$g}_3D_enabled", 0) ? 'true' : 'false';
+        if ($config3DEnabled == 'true') {
+            // Si 3D activé, on désactive la barre de navigation
+            $configNavigatorEnabled = 'false';
+            $configBarreEnabled = 'false';
+            //$configButtonsEnabled = 'false';
+        }
+        $config3DAlpha = (int)$eqLogic->getConfiguration("graph{$g}_3D_alpha", 15);
+        $config3DBeta = (int)$eqLogic->getConfiguration("graph{$g}_3D_beta", 15);
+        $config3DDepth = (int)$eqLogic->getConfiguration("graph{$g}_3D_depth", 25);
+        $config3DViewDistance = (int)$eqLogic->getConfiguration("graph{$g}_3D_view_distance", 0);
+        $configZoomAxeX = $eqLogic->getConfiguration("graph{$g}_zoom_axe_x", 1) ? true : false;
+        $configZoomAxeY = $eqLogic->getConfiguration("graph{$g}_zoom_axe_y", 1) ? true : false;
+        if ($configZoomAxeX && $configZoomAxeY) {
+            $zoomType = 'xy';
+        } elseif ($configZoomAxeX) {
+            $zoomType = 'x';
+        } elseif ($configZoomAxeY) {
+            $zoomType = 'y';
+        } else {
+            $zoomType = 'none';
+        }
 
         //log::add(__CLASS__, 'debug', "Equipment: '{$nameEqpmnt}' Graph {$g}: configbarre {$configBarreEnabled} confignavigator {$configNavigatorEnabled} configbuttons {$configButtonsEnabled}");
         
@@ -303,44 +346,44 @@ public function toHtml($_version = 'dashboard', $eqLogic = null) {
                 $startTime = ($global) ? date("Y-m-d H:i:s", strtotime($dateDebutGraph1date)) : date("Y-m-d H:i:s", strtotime($dateDebutGraph));
                 $endTime = date("Y-m-d H:i:s", time());
                 $actualisation = true;
-                log::add(__CLASS__, 'debug', "Equipment: '{$nameEqpmnt}' Graph {$g}: Using date for start time calculation. Start time: {$startTime} End time: now");
+                log::add(__CLASS__, 'debug', "Equipment: '{$nameEqpmnt}' Graph {$g}: Using date for start time calculation. Start time: {$startTime} End time: now ({$endTime})");
                 break;
             case 'nbJours':
                 $delai = ($global) ? $delaiGraph : intval($eqLogic->getConfiguration("delai_histo_graph{$g}"));
                 $startTime = date("Y-m-d H:i:s", time() - $delai * 24 * 60 * 60);
                 $endTime = date("Y-m-d H:i:s", time());
                 $actualisation = true;
-                log::add(__CLASS__, 'debug', "Equipment: '{$nameEqpmnt}' Graph {$g}: Using delay of {$delai} days for start time calculation. Start time: {$startTime} End time: now");
+                log::add(__CLASS__, 'debug', "Equipment: '{$nameEqpmnt}' Graph {$g}: Using delay of {$delai} days for start time calculation. Start time: {$startTime} End time: now ({$endTime})");
                 break;
             case 'dDay':
                 $startTime = date("Y-m-d 00:00:00", time());
                 $endTime = date("Y-m-d H:i:s", time());
                 $actualisation = true;
-                log::add(__CLASS__, 'debug', "Equipment: '{$nameEqpmnt}' Graph {$g}: Using today for start time calculation. Start time: {$startTime} End time: now");
+                log::add(__CLASS__, 'debug', "Equipment: '{$nameEqpmnt}' Graph {$g}: Using today for start time calculation. Start time: {$startTime} End time: now ({$endTime})");
                 break;
             case 'dWeek':
                 $startTime = date("Y-m-d 00:00:00", strtotime('monday this week'));
                 $endTime = date("Y-m-d H:i:s", time());
                 $actualisation = true;
-                log::add(__CLASS__, 'debug', "Equipment: '{$nameEqpmnt}' Graph {$g}: Using this week for start time calculation. Start time: {$startTime} End time: now");
+                log::add(__CLASS__, 'debug', "Equipment: '{$nameEqpmnt}' Graph {$g}: Using this week for start time calculation. Start time: {$startTime} End time: now ({$endTime})");
                 break;
             case 'dMonth':
                 $startTime = date("Y-m-01 00:00:00", time());
                 $endTime = date("Y-m-d H:i:s", time());
                 $actualisation = true;
-                log::add(__CLASS__, 'debug', "Equipment: '{$nameEqpmnt}' Graph {$g}: Using this month for start time calculation. Start time: {$startTime} End time: now");
+                log::add(__CLASS__, 'debug', "Equipment: '{$nameEqpmnt}' Graph {$g}: Using this month for start time calculation. Start time: {$startTime} End time: now ({$endTime})");
                 break;
             case 'dYear':
                 $startTime = date("Y-01-01 00:00:00", time());
                 $endTime = date("Y-m-d H:i:s", time());
                 $actualisation = true;
-                log::add(__CLASS__, 'debug', "Equipment: '{$nameEqpmnt}' Graph {$g}: Using this year for start time calculation. Start time: {$startTime} End time: now");
+                log::add(__CLASS__, 'debug', "Equipment: '{$nameEqpmnt}' Graph {$g}: Using this year for start time calculation. Start time: {$startTime} End time: now ({$endTime})");
                 break;
             case 'dAll':
                 $startTime = date("1970-01-01 00:00:00");
                 $endTime = date("Y-m-d H:i:s", time());
                 $actualisation = true;
-                log::add(__CLASS__, 'debug', "Equipment: '{$nameEqpmnt}' Graph {$g}: Using all data for start time calculation. Start time: {$startTime} End time: now");
+                log::add(__CLASS__, 'debug', "Equipment: '{$nameEqpmnt}' Graph {$g}: Using all data for start time calculation. Start time: {$startTime} End time: now ({$endTime})");
                 break;
             default:
         }
@@ -479,7 +522,7 @@ public function toHtml($_version = 'dashboard', $eqLogic = null) {
 
             $nbSeries += 1;
             
-            log::add(__CLASS__, 'debug', "Equipment: '{$nameEqpmnt}' Graph {$g} Curve {$i}: Processing with command {$cmdGraphe}, name {$indexNom}, compare={$compareType} and first={$first}");
+            //log::add(__CLASS__, 'debug', "Equipment: '{$nameEqpmnt}' Graph {$g} Curve {$i}: Processing with command {$cmdGraphe}, name {$indexNom}, compare={$compareType} and first={$first}");
 
 
             $cmd = cmd::byId(str_replace('#', '', $cmdGraphe));
@@ -742,7 +785,7 @@ public function toHtml($_version = 'dashboard', $eqLogic = null) {
             }
 
             $axisIndex = $unitToAxis[$unite] ?? 0;
-            log::add(__CLASS__, 'debug', "Equipment: '{$nameEqpmnt}' Graph {$g} Curve {$i}: Unit '{$unite}' assigned to axis index {$axisIndex}");
+            //log::add(__CLASS__, 'debug', "Equipment: '{$nameEqpmnt}' Graph {$g} Curve {$i}: Unit '{$unite}' assigned to axis index {$axisIndex}");
 
             if ($finalCurveType === 'timeline') {
                 $dataLabelsOverlap = $eqLogic->getConfiguration("graph{$g}_dataLabels_overlaps", 0) ? 'true' : 'false';
@@ -776,7 +819,7 @@ public function toHtml($_version = 'dashboard', $eqLogic = null) {
                     $actualisation = false;
                     foreach ($recordData as $year => $data) {
                         if ($year == $currentYear) $actualisation = true; // ne pas mettre à jour si pas l'année courante
-                        log::add(__CLASS__, 'debug', "Equipment: '{$nameEqpmnt}' Graph {$g} i= {$i} nbSeries= {$nbSeries} year= {$year} actualisation= " . ($actualisation ? 'true' : 'false'));
+                        //log::add(__CLASS__, 'debug', "Equipment: '{$nameEqpmnt}' Graph {$g} i= {$i} nbSeries= {$nbSeries} year= {$year} actualisation= " . ($actualisation ? 'true' : 'false'));
                         if ($rolling){
                             $years = $year . '.' . (substr($year,2,2) + 1);
                         } else {
@@ -794,6 +837,9 @@ public function toHtml($_version = 'dashboard', $eqLogic = null) {
                             },
                             yAxis: {$axisIndex},
                             stacking: '$stackingOption',
+                            dataGrouping: { {$dataGroupingJS}
+                                dateTimeLabelFormats: { {$dataGroupingDateTimeLabelFormatsJS} }
+                            },
                         },\n";
                     }
                     $xAxisJS .=  "
@@ -830,7 +876,7 @@ public function toHtml($_version = 'dashboard', $eqLogic = null) {
                     $nbSeries = count($recordData);
                     foreach ($recordData as $year => $data) {
                         if ($compareMonth == $currentMonth && $year == $currentYear) $actualisation = true; // ne pas mettre à jour si pas le mois courant de l'année courante
-                        log::add(__CLASS__, 'debug', "Equipment: '{$nameEqpmnt}' Graph {$g} i= {$i} nbSeries= {$nbSeries} comparemonth: {$compareMonth} currentmonth= {$currentMonth} year= {$year} actualisation= " . ($actualisation ? 'true' : 'false'));
+                        //log::add(__CLASS__, 'debug', "Equipment: '{$nameEqpmnt}' Graph {$g} i= {$i} nbSeries= {$nbSeries} comparemonth: {$compareMonth} currentmonth= {$currentMonth} year= {$year} actualisation= " . ($actualisation ? 'true' : 'false'));
                         $seriesJS .= "{
                             name: " . json_encode($indexNom . " - {$year}") . ",
                             borderColor: " . json_encode($color) . ",
@@ -842,6 +888,9 @@ public function toHtml($_version = 'dashboard', $eqLogic = null) {
                             },
                             yAxis: {$axisIndex},
                             stacking: '$stackingOption',
+                            dataGrouping: { {$dataGroupingJS}
+                                dateTimeLabelFormats: { {$dataGroupingDateTimeLabelFormatsJS} }
+                            },
                         },\n";
                     }
                     $xDateFormatJS = "%d %B - %Hh%M";
@@ -908,7 +957,7 @@ public function toHtml($_version = 'dashboard', $eqLogic = null) {
 
             }
             
-            log::add(__CLASS__, 'debug', "Equipment: '{$nameEqpmnt}' Graph {$g} i= {$i} nbSeries= {$nbSeries} year= {$year} cmdId= {$cmdId} actualisation= " . ($actualisation ? 'true' : 'false'));
+            //log::add(__CLASS__, 'debug', "Equipment: '{$nameEqpmnt}' Graph {$g} i= {$i} nbSeries= {$nbSeries} year= {$year} cmdId= {$cmdId} actualisation= " . ($actualisation ? 'true' : 'false'));
             if ($cmdId and $actualisation) {
                 //log::add(__CLASS__, 'debug', "Equipment: '{$nameEqpmnt}' Graph {$g} i= {$i} nbSeries= {$nbSeries} graphtype= {$graphType}");
                 //log::add(__CLASS__, 'debug', "{$graphType} data: " . json_encode($listeHisto));
@@ -983,7 +1032,7 @@ public function toHtml($_version = 'dashboard', $eqLogic = null) {
                     if ('{$cmdId}' !== '') {
                         jeedom.cmd.addUpdateFunction('{$cmdId}', function(_options) {
                             const dateLocaleMs = Math.floor(new Date().getTime()/1000) * 1000;
-                            let currentRawValue = parseFloat(_options.display_value) * {$coef};
+                            let currentRawValue = parseFloat(_options.value) * {$coef};
 
                             const variation = ({$var} === true || {$var} === 'true' || {$var} === 1 || {$var} === '1');
 
@@ -1005,7 +1054,9 @@ public function toHtml($_version = 'dashboard', $eqLogic = null) {
 
                                 // Ajout du point
                                 series.addPoint([dateLocaleMs, yToAdd], true, false, true);
-
+                                console.log('Points nouveaux : ' + yToAdd + ' Série : ' + {$nbSeries} + ' _options.unit: ' + _options.unit);
+                                console.log(_options);
+                                
                                 // Sauvegarde la valeur brute pour la prochaine mise à jour
                                 if (!series.userOptions) series.userOptions = {};
                                 series.userOptions.lastRawValue = currentRawValue;
@@ -1028,7 +1079,7 @@ public function toHtml($_version = 'dashboard', $eqLogic = null) {
             {$buttonJS},
             buttonPosition: {
                 x: {$xRangeSelectorButtonPosition},
-                y: -30
+                y: {$yRangeSelectorButtonPosition},
             }
         }";
 
@@ -1044,6 +1095,7 @@ public function toHtml($_version = 'dashboard', $eqLogic = null) {
         if ($isTimeline) {
             $headerFormatJS = ''; // pas de header pour les timelines
             $graphType = 'timeline';
+            $xAxisNavigatorJS = '';
             if ($inverted == 'true') {
                 $xAxisJS .=  "
                     dateTimeLabelFormats: {
@@ -1076,16 +1128,22 @@ public function toHtml($_version = 'dashboard', $eqLogic = null) {
         window.chart_g{$g}_id{$eqLogic->getId()} = Highcharts.{$chartOrStock}('{$containerId}', {
             chart: {
                 type: '$graphType',
+                zooming: {
+                    mouseWheel: true,
+                    type: '{$zoomType}',
+                },                
                 backgroundColor: 'transparent',
                 plotBackgroundColor: {$plotBgCode},
                 spacing: [10, 0, 10, 0],
                 inverted: {$inverted},
                 options3d: {
-                    enabled: false,
-                    alpha: 15,
-                    beta: 15,
-                    depth: 50,
-                    viewDistance: 25
+                    enabled: {$config3DEnabled},
+                    alpha: {$config3DAlpha},
+                    beta: {$config3DBeta},
+                    depth: {$config3DDepth},
+                    viewDistance: {$config3DViewDistance},
+                    fitToPlot: true,
+                    axisLabelPosition: 'auto'
                 }                
             },
             exporting: {
@@ -1189,7 +1247,7 @@ class jeeHistoGraphCmd extends cmd {
     $eqLogic = $this->getEqLogic(); //récupère l'éqlogic de la commande $this
     switch ($this->getLogicalId()) { //vérifie le logicalid de la commande      
       case 'refresh': // LogicalId de la commande rafraîchir
-        //log::add('jeeHistoGraph', 'debug', __('Exécution de la commande refresh pour l\'équipement ' . $eqLogic->getName(), __FILE__));
+        log::add('jeeHistoGraph', 'debug', __('Exécution de la commande refresh pour l\'équipement ' . $eqLogic->getName(), __FILE__));
         jeeHistoGraph::rfresh($eqLogic);
       break;
       default: